% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.Date.ts}
\alias{as.Date.ts}
\title{Convert}
\usage{
\method{as.Date}{ts}(x, ...)
}
\arguments{
\item{x}{A \code{ts} object. The \code{tsp} attribute is assumed to have
be sampled at the daily frequency with integer parts representing years.}

\item{...}{Not used}
}
\value{
A \code{Date} vector
}
\description{
This function assumes that the data in the \code{ts} object are
sampled daily. There case when \code{frequency = 12} is treated as
monthly frequency, and \code{frequency = 4} is treated as
quarterly frequency.
}

