% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\docType{data}
\name{capm}
\alias{capm}
\title{Monthly Returns on Stocks and Risk Free Asset for a CAPM model}
\format{A data frame}
\source{
\url{https://www.jstatsoft.org/index.php/jss/article/downloadSuppFile/v041i04/P.txt}
}
\usage{
capm
}
\description{
The data consist of monthly returns on four stocks (Mobil, IBM, Weyer, and Citicorp) from January 1978 to December 1987.
It also includes the interest rate on a 30-day Treasury Bill as a proxy for the risk-free asset, and
value-weighted average returns for all stocks listed on the New York and American Stock Exchanges as a proxy for
market returns.
}
\references{
Petris, G., & Petrone, S. (2011). \dQuote{State Space Models in R.} \emph{Journal of Statistical Software}, 41(4), 1 - 25.
   \href{http://dx.doi.org/10.18637/jss.v041.i04}{doi:10.18637/jss.v041.i04}
}
\keyword{datasets}

