% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parparser.R
\name{create_parnames}
\alias{create_parnames}
\alias{create_stan_parnames}
\title{Create Stan parameter names}
\usage{
create_parnames(x, d, pre = ".", sep = ".", post = "", colmajor = TRUE)

create_stan_parnames(x, d)
}
\arguments{
\item{x}{\code{character} Parameter name.}

\item{d}{\code{integer} Dimension of the array.}

\item{pre}{\code{character} String to put before indices.}

\item{sep}{\code{character} String used to seperate indices.}

\item{post}{\code{character} String to put after indices.}

\item{colmajor}{\code{logical} Order indexes by column (\code{TRUE}) or row (\code{FALSE}).}
}
\value{
\code{character} vector of flat parameter names.
}
\description{
Given a parameter name and a matrix of index values, generate
names for the unlisted parameters.
}
\examples{
create_stan_parnames("alpha", c(1, 2))
create_parnames("alpha", c(1, 2), "_", "_", "")
}

