% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisas.R
\docType{data}
\name{divisas}
\alias{divisas}
\title{Exchange Rates}
\format{A data frame with 29580 observations and 3 variables.}
\source{
\url{https://www.jstatsoft.org/index.php/jss/article/downloadSuppFile/v039i02/divisas.rda}
}
\usage{
divisas
}
\description{
Data consists of N = 2465 daily observations of the exchange rate
(against the US dollar, USD) of twelve currencies for the period
from 1988-01-04 to 1997-08-15.
}
\references{
Tusell, F. (2011). \dQuote{Kalman Filtering in R}. \emph{Journal of Statistical Software}, 39(2), 1 - 27. \href{https://dx.doi.org/10.18637/jss.v039.i02}{doi:10.18637/jss.v039.i02}
}
\keyword{datasets}

