% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{extract_param_from_ssm_filter}
\alias{extract_param_from_ssm_filter}
\title{Extract parameters from Stan State Space results}
\usage{
extract_param_from_ssm_filter(x, m, p, params = NULL)
}
\arguments{
\item{x}{Array containing results from \code{ssm_filter} Stan function. This
is usually extracted from a \code{stanfit} object.}

\item{m}{Number of states}

\item{p}{Dimension of observation vector}

\item{params}{Parameters to extract from the filter. One of
\code{"v"}, \code{"Finv"}, \code{"K"}, \code{"a"}, \code{"P"}.
If \code{NULL}, all values are extracted.}
}
\value{
A named \code{list} of \code{array} objects, one for each parameter.
}
\description{
Extract results from the output of the \code{ssm_filter} function
in stan.
}

