% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parparser.R
\name{parse_parnames}
\alias{parse_parnames}
\alias{parse_stan_parnames}
\title{Parse Stan parameter names}
\usage{
parse_parnames(x, pre = "\\\\W+", sep = "\\\\W+", post = "\\\\W+")

parse_stan_parnames(x)
}
\arguments{
\item{x}{\code{character} vector with flat parameter names.}

\item{pre}{\code{character} Pattern between parameter name and indices.}

\item{sep}{\code{character} Pattern seperating each index.}

\item{post}{\code{character} Pattern following the indices.}
}
\value{
A data frame with columns: `\code{paramname}`, `\code{parameter}`, and
 `\code{dim_1}`, `\code{dim_2}`, ....
}
\description{
Functions that parse a vector of flatted parameter names
and returns a data frame with the parameter names and indexes.
}
\examples{
parse_stan_parnames(c("alpha", "beta[1]", "gamma[1,1]", "gamma[1, 2]"))
parse_parnames(c("beta.1.1", "beta.1.2"))
}

