% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm_extract}
\alias{ssm_extract}
\title{Extract parameters from Stan State Space Model results}
\usage{
ssm_extract(x, m, p, q = m, type = c("filter", "filter_states",
  "smooth_state", "smooth_eps", "smooth_eta", "sim_rng"), params = NULL)
}
\arguments{
\item{x}{An array containing results from one of the "ssm" Stan functions:
\code{ssm_filter}, \code{ssm_filter_states}, \code{ssm_smooth_eta},
\code{ssm_smooth_eps}, \code{ssm_smooth_state}, or \code{ssm_sim_rng}.
This array is usually extracted from a \code{stanfit} object using
\code{\link[rstan]{extract}}.}

\item{m}{Dimension of the state vector}

\item{p}{Dimension of observation vector}

\item{q}{Dimension of the state disturbance vector}

\item{type}{Character. The function which generated the vector.}

\item{params}{Character. Then names of arameters to extract from the filter. See @details
for the available parameters for each type.
If \code{NULL}, then all parameters for that type are extracted.}
}
\value{
A named \code{list} of \code{array} objects, one for each parameter.
}
\description{
Extract results from the output of the \code{ssm_filter} function
in stan.
}
\details{
The parameters returned by each \code{type} are:
\describe{
\item{\code{"filter"}}{\code{"v"}, \code{"Finv"}, \code{"K"}, \code{"a"}, \code{"P"}}
\item{\code{"filter_states"}}{\code{"a"}, \code{"P"}}
\item{\code{"smooth_state"}}{\code{"alpha"}, \code{"V"}}
\item{\code{"smooth_eta"}}{\code{"mean"}, \code{"var"}}
\item{\code{"smooth_eps"}}{\code{"mean"}, \code{"var"}}
}
}

