% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm_process_na}
\alias{ssm_process_na}
\title{Process missing values for Stan functions}
\usage{
ssm_process_na(x)
}
\arguments{
\item{x}{A matrix possibly containing missing values.}
}
\value{
A list with elements:
\describe{
\item{n}{an integer array of the number of non-missing values}
\item{idx}{and an 2-dimensional integer array of the indexes of the non-missing values (padded with zeros)}
}
}
\description{
Process a matrix with missing values into the inputs that the
Stan functions require for missing values.
}

