% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{ssm_stan_include_path}
\alias{ssm_available_models}
\alias{ssm_stan_include_path}
\alias{ssm_stan_model}
\alias{ssm_stan_model_path}
\title{Stan Paths and Files}
\usage{
ssm_stan_include_path()

ssm_stan_model_path()

ssm_available_models()

ssm_stan_model(file, ...)
}
\arguments{
\item{file}{The name of the Stan model to run}

\item{...}{Passed to \code{\link[rstan]{stan}}}
}
\description{
\code{ssm_stan_include_path} returns the path to
the directory with the Stan user-defined functions provided
by this package. This is used with the \code{isystem}
argument in \code{\link[rstan]{stanc_builder}}.
}
\details{
\code{ssm_stan_models_path} returns the path to
the directory containing the models provided by this package.

\code{ssm_available_models} lists the names of the stan models
provided by this package.
\code{ssm_stan_model} runs a Stan model included in this package.
This function is a wrapper for \code{\link[rstan]{stan}}, that
sets the correct path to the model and includes the functions provided
in this package.
}

