% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stan_model_view_code}
\alias{stan_model_view_code}
\title{View the model code of compiled Stan model}
\usage{
stan_model_view_code(mod, start = NULL, end = NULL)
}
\arguments{
\item{mod}{A Stan model}

\item{start}{Start line}

\item{end}{End line}
}
\description{
This can be useful when debugging and the model contains include statements
that makes it difficult to find the problematic line in the original file.
}

