% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm.R
\name{tidy_stan_summary}
\alias{tidy_stan_summary}
\title{Tidy the results of Stanfit summary objects}
\usage{
tidy_stan_summary(x)
}
\arguments{
\item{x}{The list returned by calling \code{summary} on a \code{Stanfit} object.}
}
\value{
A list with two elements,
\describe{
\item{all}{A data frame with summary statistics for all the chains combined.}
\item{chains}{A data frame with summary statistics for each chain.}
}
}
\description{
Tidy the results of Stanfit summary objects
}

