% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unit_lower_tri}
\alias{eye}
\alias{selection_matrix}
\alias{unit_lower_tri}
\alias{unit_upper_tri}
\title{Upper and lower unit triangle matrices}
\usage{
unit_lower_tri(m, n = m, diag = TRUE)

unit_upper_tri(m, n = m, diag = TRUE)

selection_matrix(m, idx, rows = TRUE)

eye(m, n = m, k = 0)
}
\arguments{
\item{m}{integer. Number of rows in the matrix}

\item{n}{integer. Number of columns in the matrix}

\item{diag}{logical. Include diagonal?}

\item{idx}{integer. Indexes of rows (if \code{rows = TRUE}),
or columns (if \code{rows = FALSE}) to select.}

\item{rows}{logical. If \code{TRUE}, then the selection matrix
selects rows, if false, then it selects columns.}

\item{k}{Index of the diagonal. 0 is the main diagonal, positive numbers for
upper diagonals, negative numbers for lower diagonals.}
}
\value{
For \code{unit_lower_tri} and \code{unit_upper_tri} a
  matrix of dimensions \code{c(m, n)}. For \code{selection_matrix}
  a matrix with dimensions \code{m, length(idx)} if \code{rows = TRUE},
  else a matrix with dimensions \code{c(length(idx), m}
}
\description{
Some useful functions for generating matrices commonly encountered
in state space models.
}
\section{Functions}{
\itemize{
\item \code{unit_upper_tri}: Generate an unit upper triangular matrix.

\item \code{selection_matrix}: Generate a selection matrix.

\item \code{eye}: Generate matrix with ones on the diagonal
}}

