% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_to_symmat}
\alias{vector_to_symmat}
\title{Vector to symmetric matrix}
\usage{
vector_to_symmat(x)
}
\arguments{
\item{x}{vector with n (n + 1) / 2 elements.}
}
\value{
A n x n symmetric matrix

This fills in the matrix assuming that the elements
are from the lower triangular part of the symmetric
matrix, and were filled in column-wise.
}
\description{
Vector to symmetric matrix
}

